refactor(OutputMessage, PrinterConstants, Printer): 여러 파일에 분산된 Output 관련 상수를 통합
